pyproject.toml
[tool.poetry]
name = "fastapi-odmantic-example"
version = "0.1.0"
description = ""
authors = ["jakee <jakee.indapanya@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.1"
odmantic = "^0.9.0"
uvicorn = "^0.19.0"
python-dotenv = "^0.21.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.5"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "app.cmd.run:main"
